Fix linter

+++ internal/domain/product/repository.go
@@ -8,43 +8,43 @@ import (
 
 //go:generate mockgen -source $GOFILE -destination "repository_mock.go" -package "product"
 
-// Repository
+// Repository represents product repository.
 type Repository interface {
 	CreateTx(ctx context.Context, props *Props) (*Product, error)
 	Get(ctx context.Context, filter *Filter) (*Product, error)
 	Query(ctx context.Context, criteria *QueryCriteria) (*QueryResult, error)
 }
 
-// QueryCriteria
+// QueryCriteria represents criteria for product query.
 type QueryCriteria struct {
 	Filter     Filter
 	Ordering   Ordering
 	Pagination Pagination
 }
 
-// QueryResult
+// QueryResult represents a result for products query.
 type QueryResult struct {
 	Items Products
 	Count int
 }
 
-// Filter
+// Filter represents product filter.
 type Filter struct {
 	ID         IDFilter
 	CategoryID category.IDFilter
 }
 
-// IDFilter
+// IDFilter represents ID filter.
 type IDFilter struct {
 	Eq  IDs
 	Neq IDs
 }
 
-// Ordering
+// Ordering represents order criteria.
 type Ordering struct{}
 
-// Pagination
+// Pagination is a pagination parameters.
 type Pagination struct {
-	Limit  int
-	Offset int
+	Limit  uint64
+	Offset uint64
 }
